删除adminPage的复制按钮，修复admin页面加载bug
