+++ .env
@@ -1,3 +1,3 @@
 # VITE_API_URL=http://localhost:8000
 VITE_API_URL=http://8.130.141.190:8000
-VITE_FRONTEND_URL=http://8.130.141.190
+VITE_WEBSITE_UTL = http://8.130.141.190
