+++ nginx_conf_and_Dockerfile/nginx.conf
@@ -1,0 +1,28 @@
+server {
+    listen 80;
+    server_name 8.130.141.190;  # 替换成您的实际域名，如果是本地开发可以设置为localhost
+
+    root /usr/share/nginx/html;  # 这个路径对应Docker容器内部的Vue项目静态资源根目录
+
+    location / {
+        index index.html;
+        try_files $uri $uri/ /index.html;  # 对Vue.js单页应用很重要的部分，当找不到请求的文件时，返回/index.html
+        autoindex off;
+    }
+
+    # 可选配置，如果您的Vue应用中有API请求需要代理到后端服务器
+    # location /api/ {
+    #     proxy_pass http://backend-server:port;  # 替换为您的后端服务器地址和端口
+    #     proxy_set_header Host $host;
+    #     proxy_set_header X-Real-IP $remote_addr;
+    #     proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+    #     proxy_set_header X-Forwarded-Proto $scheme;
+    # }
+
+    error_page 404 /404.html;  # 如果想自定义404错误页面，确保在静态资源目录下存在404.html文件
+    error_page 500 502 503 504 /50x.html;
+
+    location = /50x.html {
+        root /usr/share/nginx/html;
+    }
+}
